// frontend/src/api/models/GetUserInfo200Response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Assignment 2 - User Login Service API
 * Assignment 2 用户登录服务 MVP - API规范文档  这是一个基于 Go Gin 框架构建的用户认证服务，提供登录和用户信息获取功能。  ## 特性 - JWT token 认证 - Bcrypt 密码加密 - RESTful API 设计 - CORS 支持  ## 认证流程 1. 用户通过 `/user/login` 接口进行身份验证 2. 服务器验证用户名和密码，返回 JWT token 3. 客户端在后续请求中通过 `Authorization: Bearer <token>` 头部携带 token 4. 服务器验证 token 有效性，提供相应的服务 
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: developer@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface GetUserInfo200Response
 */
export interface GetUserInfo200Response {
    /**
     * 用户昵称
     * @type {string}
     * @memberof GetUserInfo200Response
     */
    nickName?: string;
}

/**
 * Check if a given object implements the GetUserInfo200Response interface.
 */
export function instanceOfGetUserInfo200Response(value: object): value is GetUserInfo200Response {
    return true;
}

export function GetUserInfo200ResponseFromJSON(json: any): GetUserInfo200Response {
    return GetUserInfo200ResponseFromJSONTyped(json, false);
}

export function GetUserInfo200ResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetUserInfo200Response {
    if (json == null) {
        return json;
    }
    return {
        
        'nickName': json['nick_name'] == null ? undefined : json['nick_name'],
    };
}

export function GetUserInfo200ResponseToJSON(json: any): GetUserInfo200Response {
    return GetUserInfo200ResponseToJSONTyped(json, false);
}

export function GetUserInfo200ResponseToJSONTyped(value?: GetUserInfo200Response | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'nick_name': value['nickName'],
    };
}